Fix weird exception: "PYZ entry 'encodings' (0j) is not a valid code object".
The actual problem is that marshal.loads() is feed a compressed stream.
This happens because zlib could not be imported because of a packaging/extraction
bug. Thus, instead of raising this weird exception, we prefer to raise
a straightforward RuntimeError "cannot import zlib", so that at least the
cause is clear.

The underlying problem of course must be fixed elsewhere...