Improve the code to generate the temporary directory.
On Windows, there is no race-free library function, so we use
_tempnam which ensure uniqueness at the time of call, and then
create the directory afterwards. If it fails before of a race,
we retry a few times. Anyway, we keep the pid in the name prefix,
so that it's hard to foresee a race.

On Linux and Mac OSX, we use mkdtemp() which is totally race-free
and creates the directory for us. This part still needs testing
but should work.