Complete runtime support for .egg-files.
Now Build.py automatically adds a runtime hook whenever
a .egg file is packaged. This hook adds the .egg to sys.path
at startup, so that it is picked up by the import hooks.

In one-dir mode, .eggs files are packaged into a "eggs" subdirectory.
In one-file mode, .eggs files are within the executable, and
extracted at startup in the temporary directory.

I removed the old runtime hook by Hartmut which decompressed .eggs
to disk because it is not required anymore since we directly
use zipimport to access .eggs.