Fixed some ifdefs which belong to *Windows* windowed bootloader only.
Windows definitions for the windowed bootloader (e.g. WinMain) were
located in a specific windows/winmain.c file before the bootloader
refactoring that happened in the multiprocess branch; they had been
erroneously imported 'as-is', resulting in some ifdef'd hunks that
were enabled while building the windowed bootloader on all platforms,
but that actually only belonged to the *Windows* windowed bootloader.