Rework the 32-bit/64-bit warning on Darwin.
It looks like there is not a single way to switch to the 32-bit version
of Python: you must either set an environment variable or use the 'arch'
command, depending on whether you are using the generic /usr/bin/python
or a version-specific wrapper (eg. python2.6).

So we rework the warning message to apply both suggestions at the same
time, so that it will work by just copy & pasting without understanding
the whole mess.