Use git-revision in version only if running in the development-checkout.
This fixes: If PyInstaller is installed from some sdist into a virtualenv
below the development-checkout, this PyInstaller does always show the
git-revision of the checkout, not the one of the sdist. Now you still
will not get the revition of the sdist, but an enpty string, so it is
distinguishable at least.

Still todo: put the git-revision into the sdist/bdist somehow.