Fixed bootloader on AIX for loading libpython.
On AIX the define _ALL_SOURCE must be set in order for the dlopen()
flag RTLD_MEMBER to be defined. This flag is necessary to load a shared
object from an archive.

Also, the special format /path/to/archive.a(archive.so) must be used
to reference the shared object.

This fixes ticket 1044 'AIX support broken in version 2.1'.