WIP - trying to fix MacOSX codesigning problem
MacOSX doesn't like directories that end in .digit inside the MacOS folder, e.g. myapp.app/Contents/sitepackages/gst-0.1. So try to put files in MyApp.app/Contents/Resources. Files are going in the right place now, but my game dies when loading gstreamer. Not sure if that's another problem or not.