new hook - pycparser
Packaging without this hook generates an executable that seems to work.

However, on running, pycparser generates the lextab.py and yacctab.py
modules. This may fail in cases where the executable has no permissions
to write wherever it wants to write: which seems to be the same directory
as the remaining packaged binaries.

Having no knowledge of pycparser itself (I'm using it as a dependency
for cffi which, in turn, is a dependency for cryptography) I'm not sure
if this is the optimal approach.