adding a test for pycparser hook
pycparser needs two modules -- lextab.py and yacctab.py -- which
it generates at runtime if they cannot be imported. Those modules
are written to the current working directory for which the running
process may not have write permissios, leading to a runtime
exception.

This test checks to see if any files were generated or not. If not,
it means the hook did its job correctly by defining lextab.py and
yacctab.py as hidden imports.