Fix TOC's minus operator ignoring normalized filenames.
Previously, subtracting an empty TOC from another one (as in `a.binaries = a.binaries - TOC([])`) would return a TOC that is missing tuples whose filename != its normalized filename. Specifically, this causes the VC90 runtime libs to be dropped on win32.

This fix to `TOC.__sub__` considers that filenames in self.fltr are normalized, but filenames in the tuples are not.