Add background option.
The background option enables OSX apps to avoid displaying an icon in the dock when it doesn't make since to do so. Setting background to True in EXE uses a new background loader that doesn't have the window bootloader's icon properties but is otherwise identical. The specfile generator also has a background option that is set to False by default. The background option should only be used with console=False.