"_tkinter" module hook fixed.
The prior version of this hook erroneously appended to the "mod.datas" list.
Since the Analysis.assemble() method ignored that list, this hook was equivalent
to a noop. This corrects that by calling the newly minted mod.add_data() method
instead. For readability, this also improves hook docstrings and function names.

Conflicts:
	PyInstaller/hooks/hook-_tkinter.py