Make runtests call archive_viewer instead of running it as a subprocess.
This a) speeds up tests, b) avoids encoding and decoding in-output and
c) avoids eval().

This also avoids calling a different script when running in a virtual env,
because if we are in a virtual env, sys.path is not expanded (see top of
file) and thus the imported archive_viewer is the same as the one used
by the pyi-archive_viewer installed in the virtual env.