Clean up and fix the way generated test-case-executables are run.
Decouple finding the exe and actually run it. This also is a good
preperation for better unittest-messages (see commits comming soon) and
eliminated searching for the executable twice.

This also fixes that test_exe() did not find "all generaed executables"
(as the doc-strings says), but only those starting with the full
test-name.