Add skipping test-cases known to fail and option to run them anyway.
This commit adds a list of these test-cases (with reason), makes runtest
skip them by default and adds an command-line option `--known-fails`
to enforce running them.

Running test-cases which are failing due to a known reason (which can
not be solved easily) is confusing for the developers. Also it renders
continuous-integration testing unusable because it returns misleading
results.