Changed Analysis to not write `self.pure['code']` to out00-Analysis.toc
This was causing the tocfile to never be readable and the Analysis step always reports `INFO: Building because out00-Analysis.toc missing or bad`. This change makes the tocfile readable again and the Analysis step can be skipped as usual if nothing else changed.

In `Analysis.check_guts()`, the `pure` from the tocfile is correctly placed in `self.pure['toc']` already, so this change is all that is needed.