Don't embed full pathname of PYZ in compiled python code.
ZlibArchive is responsible for finding files specified in its TOC and compiling the python source to bytecode. If both .py and .py[co] files are present, it prefers to read the .py file and recompile it.

It compiles the source using the `compile()` builtin which takes as arguments the source text plus an "effective filename" to embed into the source for traceback purposes. This change simply uses the basename of the .pyz file to compute the effective filename, instead of the full path to the .pyz file.

This change has no effect for .py[co] files that have no corresponding .py file present.