Suppress Windows Error Reporting dialogs while calling Py_Initialize
Py_Initialize calls abort() to terminate the entire application. This is very rude, especially for embedded pythons.

abort() on Windows triggers the Windows Error Reporting Dialog, which will pause the test suite until it is dismissed. This patch
disables the dialog during the call to Py_Initialize for debug builds of the bootloader.