Remove useless unsetenv("PYTHONHOME").
Doing this is no longer required, see History for Python 1.5.2a1:

"The code that initializes sys.path now calls Py_GetPythonHome()
instead of getenv("PYTHONHOME").  This, together with the new API
Py_SetPythonHome(), makes it easier for embedding applications to
change the notion of Python's "home" directory (where the libraries
etc. are sought)."