Fix global defined in header.
Define global in pyi_python27_compat.c, put 'extern' declaration in header.

When is_py2 is defined in a header, it is also defined in every .c file that includes it. Instead, define is_py2 in only one .c file and put an 'extern' declaration in the header, so other .c files get linked to the global in python27_compat.c