`[save,load]_py_data_struct` now opens the tocfile using an explicit UTF8 encoding.
On Windows, the default encoding is a codepage so some data can't be written.

On Python 2, the 'unicode' strings are formatted with 'backslashescape' so only ASCII is written to the file anyway, but Python 3's pprint attempts to encode the strings using the I/O encoding of the fileobj given to it. This raises an encoding error if the I/O encoding is not some UTF.