Cleanup and rework checking the guts (caches build-data).
This puts the code fro the head onto the feed. In the old code, the guts
has been read be `check_guts()`, but the method for reading the guts
(`get_guts()`) did some checks, too. In the new code, `__postinit__()`
is controling this process, calling `get_guts()` for reading and
`check_guts()` for checking.

Hopefully this is more easy to understand than the old code.