Add -Wimplicit-function-declaration to CFLAGS
This will give an error if we fail to include a header file for a function declaration.

This can be a serious error on Windows, because the implicit declaration returns 'int', which is only 32-bits on Windows. If the function returns a 64-bit value (such as a pointer), the high bits of the return value are undefined.