Change order of running tests: test logs prior to exe.
This seams to make more sense and should speed up the test-suite.

For many tests we have a log-test (aka .toc-test) for, the log-test
is the important one as the exe may not even don anything useful. For
other test-cases, the exe-test may not pass anyway if the log-test
fails.