Add functions to convert wchar_t and __wargv to UTF-8 and ANSI-encoded strings
Also reimplements pyi_win32_utils_[from,to]_utf8 in terms of MultiByteToWideString(CP_UTF8) and alters them to return a malloc'd buffer if the output argument is NULL.

Includes variants that try to get an MS-DOS ShortFileName for the string or argv.
Includes convenience functions to convert UTF-8 to ANSI with SFN, and to convert a UTF-8 argv back to wargv.
Includes a function analogous to `strerror` for getting a string describing a Windows error code.