Issue 1438: Add more methods to `pyi_path_executable`
These methods attempt to use `/proc` entries on systems that provide them. If that fails, it checks whether argv[0] is absolute or relative and converts it to an absolute path, resolving symlinks. If argv[0] is a plain program name, it instead searches $PATH for the program to get the absolute path.

On Windows, GetModuleFileNameW is used again, because CreateProcessW (like execl) can pass anything it wants for argv[0].