Issue #1325: Improve text handling in exec_command and exec_python
On Python 3, exec_python sets PYTHONIOENCODING and decodes using the same encoding.
On Python 2, no encoding/decoding is done - output is `str` instead of `unicode`

On Python 3, exec_command decodes using the value of the new `encoding` kwarg
On Python 2, no encoding/decoding is done - output is `str` instead of `unicode`

`encoding` is the FS encoding by default. The common case is reading filenames from the output of non-Python commands. The second most common case is reading version numbers, which are almost always ASCII.

We will revisit this if we decide to change the build to use `unicode` filenames on Python 2. This change is outrageously non-trivial, so these functions just return `str` on Py2 for now.