Assemblies: Add redirects argument to Dependencies
This is used to store any binding redirects we use while searching for assemblies, so the same redirects are applied to the built application and all binaries bundled with it.

The redirects are cached in the Analysis.toc and also stored in CONF so PKG/COLLECT can apply the redirects.