Logging: Defer initialization until __process_options is called.
This solves a problem where logging was initialized for the test suites as soon as `conftest.py` was imported, and if it was imported while pytest output capturing is not enabled, all the logging was sent to stderr instead of being captured.

Now, logging is only initialized upon a call to `__main__.run()` (or to the `run()` function of the cliutils that use the logging options).

Related to https://github.com/pytest-dev/pytest-xdist/issues/4