AppVeyor: Rebuild only missing or out-of-date wheels
The wheel cache is no longer cleared by AppVeyor when the requirements file changes. Instead, the `--find-links` option is given to `pip wheel` to allow it to find and skip any wheels that were previously built and cached. Wheels will still be rebuilt if a newer version is found on PyPI and downloaded.

In the future we may need to remove out-of-date wheels from the wheel cache. There is a 100MB limit (well, "recommendation") to the build cache - I don't know what AppVeyor will do if we exceed that limit.