Change hook-wx.lib.pubsub to include several python files as data-files.
The folders `pubsub1` and `pubsub2` do not have an `__init__.py` so the files in them are not visible as Python modules, so neither collect_submodules nor hiddenimports can find them. They are included as data files instead.

The file `autosetuppubsubv1.py` is "imported" by wxPython using `imp.find_module`, which cannot find modules in our PYZ archive. It too is included as a data file only to make it visible to `imp.find_module`.