Change matchDLLArch to check machine type of sys.executable
platform.machine() does not return the machine type of sys.executable, but the machine type of the OS, which is always 64-bit even when running 32-bit Pythons. So platform.machine() is wrong.

Using the FILE_HEADER.Machine of sys.executable should be more reliable than trying to derive the machine type from sys.architecture.