bootloader: fix PATH search for executable, fixes #1634
There was some code missing: actually checking whether a path/filename
assembled from PATH component and appname exists and is executable,
and, if that is the case, returning that path/filename as success result.

In the other case (whole PATH traversed, nothing found), return code
was fixed to failure (-1).

The caller also needed some fixing. If the PATH search is unsuccessful,
it now immediately returns failure (-1). If it was successful, it runs
pyi_path_fullpath on the result to make sure nothing strange is the returned
path/filename (like ., .., // or so).

Also fixed: when it says "Cannot get fullpath for ..." it should give the
input value of the failed pyi_path_fullpath() call, not the (undefined?)
output value.

Cosmetical change: instead of "== false", just use "!" (it is a boolean
return value).