Issue #1643: Fix including of previously excluded imports.
Check if a module should be included/excluded in ModuleGraph method
_safe_import_module() was not reliable. ModuleGraph created nodes
MissingModule() for excludedimports and thus later those excluded
imports could not be found and bundled properly. The right place
to test excluding/including module seems to be in method
_process_imports(). In this method it allows to remove excluded
module directly from the module dependency list. For modulegraph
it looks then like modules from excludedimports are not even
dependencies of any module.