Fix: Imports in byte-code not found if code contains a function.
If a module was imported as byte-code (which seams to happen in eggs,
but not in e.g. the file-system) the code contains a function, the
list of imported modules was cleared when scanning the function's
code.

I introduced this bug in ff8a9ddc184da350129534d4b1f205ba5c1dbe29

See #1581.