Fix crashes in scan_code_instruction_for_ctypes: default arguments.
scan_code_instruction_for_ctypes crashed if `ctypes.util.find_library`
is passed as an default argument to an function like here::

   def locate_library(loader=ctypes.util.find_library)

which results in a byte-code like this::

  LOAD_NAME         0 (ctypes)
  LOAD_ATTR        13 (util)
  LOAD_ATTR        14 (find_library)
  LOAD_CONST       16 (<code object locate_library ...>)
  MAKE_FUNCTION     1
  STORE_NAME       15 (locate_library)

In this case `__scan_code_instruction_for_ctypes` returned the
code-object (sic!). I solved this by now testing if the const
is a string.