Fix: namespace packages implemented as nspkg.pth-files in pathex not found.
This is done by re-searching for nspkg.pth-files after extenting the
modulegraph's path.

This error only effected Python 2.7 as for some reason (which I did no
investigate further) in Python 3 these namespace-packages have been
found via package meta-data (modulegraph._namespace_package_path()).

This is why I removed the package meta-data for some of the packages in
the related test-case.

Regarding the change in the test-case: calling site.addsitedir() is
not necessary since modulgraph already implemented appropriate
code - which just did not work when extending the path.