compat.open_file() utility function added.
A new compat.open_file() utility function simplifying file handling across both
Python 2 and 3 has been added. Under Python 3, this aliases the open() built-in;
under Python 2, this aliases the io.open() built-in, which is a backport of the
Python 3 open() built-in and hence usually preferred to the mostly obsolete
Python 2 open() built-in.