import 'queue' does not work in python 2.7
Problem appears to be that pyinstaller cannot have two modules of the
same name that differ only by lower/upper case.  The from the future
'queue' simply imports all of the 'Queue' module.  So, by my reading,
since 'queue' and 'Queue' can not coexist in a frozen app, and since
'queue' requires 'Queue', there is no way to use 'queue' module in a
frozen 2.7 app.

This commit has:

1) a pre_safe_import_module hook to alias 'queue' to 'Queue'.
2) a hook to force the import of 'Queue' when importing 'queue'.
3) a hack to the loader to load 'Queue' when 'queue' is requested.
4) a test case for this fuctionality.