Add data and binaries command line options - improvements
**Improvements to my initial pull request #1990**

I added a command line option for binaries as well as for data, and I
changed the names and descriptions of both (now `--add-data` and
`--add-binary`). Please see my comment in this code section,
`--adddata`  as an alias for `--add-data` is not that good, what do you
think?

Using a list comprehension was a good idea @htgoebel, you only missed a
`)` there :)

About the manual: I've looked through all the `.rst` files, the *Adding
Files to the Bundle* section is in `spec-files.rst`. I've added a
description, but please comment on that because I'm not that familar
with documenting code...