Fix: Correct usage of multiprocessing test.
Fixes:
- In general, the superclass' constructor should be invoked first, unless there's a compelling reason not to do so.
- More descriptive diagnostic messages were added.
- Following the example of https://docs.python.org/2/library/multiprocessing.html#the-process-class, the processes are join()ed.
- Items put into the Queue are now removed. Quoting the programming guidelines at https://docs.python.org/2/library/multiprocessing.html#programming-guidelines:

Joining processes that use queues

Bear in mind that a process that has put items in a queue will wait before terminating until all the buffered items are fed by the “feeder” thread to the underlying pipe. (The child process can call the cancel_join_thread() method of the queue to avoid this behaviour.)

This means that whenever you use a queue you need to make sure that all items which have been put on the queue will eventually be removed before the process is joined. Otherwise you cannot be sure that processes which have put items on the queue will terminate. Remember also that non-daemonic processes will be joined automatically.