Build bootloader: wscript: Build zlib if required, not if "on windows".
Whether we need to build out zlib does not depend on the platform
(is_win), but on whether it is provided by the operating system. Currently
for Windows we do not check for zlib and for other platforms it is
mandatory - so the result is the same. But we make zlib optional for
Windows if mingw somewhen provides it.

* phase configure: "use" zlib as "Z"
* build: Move Building zlib from our source out of "is_win" but
  conditionally build it, if "Z" is not available.
* Change name of target zlib to "Z" to align with configure results.