MERGE: Automatically detect PYZ files that have been included in another executable.
This adds an _EXE attribute for every PYZ that is passed to an EXE. On inclusion in
a subsequence EXE, the PYZ is not added but a reference to the first EXE in which
the PYZ was included. This is simplier than the original approach, enhancing maintainability,
because the original approach only moved some dependencies into the other EXE.

This assumes that the first EXE will not be distributed independently, which should
be a valid assumption for almost all cases in which MERGE is used. In such cases, the
first EXE would be of a larger size than required, but would still function normally.