Tests: CI: Rearrange requirements files.
Use one requiements file with environment markers for listing all libraries to
be tested. This eases maintenance, since all requirements are listed at one
place.

Unfortunately environment markers do not allow to distinguish between 32 or 64
bit systems, only the machine. But I think, this is acceptable: We had a
single package which was effected (and is now solved) and the times this was
necessary should be gone now.

* Merge platform-specific files into the formerly platform-agnostic file.
* Move developer tools into a separate requirements file to reduce size
  of installed files on CI servers.
* .travis.yml: Adopt file-names.
* appveyor.yml: Adopt file-names, remove special handling for 32 bit.
* Update tests/README.md.