Tests: unit/modulegraph: Add basic tests for _scan_code().
These are basic tests just to ensure _scan_code is working. Tests for
more cases need to be added, but should go directly into upstream.

Indeed modulegraph does not have working test-cases for this low-level
function. (There is a test, but it is marked as xfail.) Thus even obvious
errors like using an undefined variable in _scan_code() or
_scan_bytecode() are unnoticed.