Issue #2623: Remove all calls to `pywintypes.Unicode()`
This function (and its use in versioninfo.py) predate proper unicode
support in Python. Currently, it is just equivalent to
`str.decode(encoding='mbcs')`.

All of its uses coincided with calls to a `getRaw()`
function to properly encode the string as UTF-16, so this function also
handles the `mbcs` decode needed on Python 2.