bootloader: set proper LISTEN_PID in child process
if the program is actived by a systemd socket, then there are some
envs(LISTEN_PID, LISTEN_FDS) for it.

LISTEN_PID is set to the pid which is forked by systemd that is the
parent process of the bootloader.

This patch is trying to correct the env to the child pid of bootloader.
parent fds are duplicated to child process, so it's OK to modify
LISTEN_PID.

Signed-off-by: Shengjing Zhu <i@zhsj.me>