Depend: Import 'pefile' only on windows.
This fixes building the documentation at read-the-docs:
Since 835fd2e3e6786f54a292c7d1ff140e497accffa4 'pefile' was imported
unconditionally, even when running `pyinstaller --help`. And when
building the docs this very command is run to generate the input for
the man-pages. But 'pefile' is not available at read-the-docs and the
build fails.