Bootloader: libdl to be thread-safe requires libpthread
libdl has dlerror call, which has internal buffer being initialized in
the first call. If libpthread is linked, this buffer being initialized
as thread-local storage, otherwise it is just static buffer.

Since bootloader is not linked with libpthread, dlerror's buffer is
initialized as non-thread-safe static buffer during the first use of
dlopen. Afterward libpthread will be loaded (libpython is linked with
it) and libdl could be used in parallel threads (python modules use it
widely), but dlerror's buffer will stay the same - non-thread-safe as
it was initialized initialy. So it may lead to crashes with
double-free or memory corruption errors.

Similar problem is also described in the following bugreport:
Bug 18192: dlerror not thread-safe when libpthread loaded through dlopen
https://sourceware.org/bugzilla/show_bug.cgi?id=18192