doc: update runtime information
Just checking `sys.frozen` is not enough to check if a program is
running from PyInstaller - some other Python packages also seem to set
this.

Mention that in most cases `__file__` can be used from a bundled app as
one usually does. Also mention that this does not work from the main
script as it doesn't know where it lives in the module hierarchy.

Fix #2882

[ci skip]