modulegraph: Keep order of imported identifiers.
The order of imported identifiers was mixed up by using a set. While this
is not important for variables and names, it may be important for
modules (see test_import_order_1, which now passes).

The performance loss for using a list instead of a set should be
negligible and later the `fromlist` is only used for iterating over it.