Remove badges from Readme.
Beside this being some kind of a eye-catcher, there is no benefit in having
these badges. Contrary the status shown here is inconstant, misleading and of
less use.

Also we have a privacy issue: Using these badges be obey our visitors to
shields.io. And shields.io seems to track users, since it sets cookies – which
seems to have the sole use of tracking users.

Closes #2853. [skip ci]