Tests: enable parallel testing on windows
Here, we enable parallel testing on windows
by working around a bug in the pytest framework:
if the sys.stdout is passed to a subprocess, it
is randomly closed. Therefore, we buffer the
output in memory and then write it ourselves. As
the subprocess output is not really that verbose
(compared to the log output), this is not an issue.

In addition, we patch subprocess32 into psutil
on Python 2.7 so that we can call
communicate(timeout). It's a hack, but it's only
used for testing, so there should be no broader
implications.