Fix replacement of runtime search path keywords
Previously runtime search path keywords like @loader_path and
@executable_path were only properly replaced if they were followed by
a slash, for example for a value like "@executable_path/Frameworks".
If the LC_RPATH value is just "@executable_path" without slashes, it
would not be replaced, resulting in an invalid path string
computed from executable_path + "/" + "@executable_path" (the exact
string literal).
The fix is to match the keyword when it is followed by a slash,
or when it is not followed by anything (end of string).