Fix ImportError for sysconfig for 3.5.4 Conda
The CPython implementation adds additional modules for importing in sysconfig for Python 3.6 and above.

Anaconda's Python 3.5.4 backports this feature for this version. This fixes the hook by removing the Python version check since there are cases where Python 3.5 may have the additional modules. Instead, sysconfig is checked for a function that was added for the additional module to get the name.

Fixes #3105