Only update resources when necessary to keep signature valid
This fixes #2526. Pyinstaller is updating the resources for python35.dll and makes its signature invalid. This then prevents that dll from being signed with the 'signtool' utility (which then can have more consequences, like the inability to publish the app to a repository where signing is mandatory).

In this commit, the resource update is skipped if it is not necessary, i.e. it is only performed if there have been som redirections or if the dll has been made private.