Hooks: hook-usb: Resolve library name reported by usb.backend.
hook-usb.py "asks" pyusb which library to pack by inspecting
usb.backend. This only results in a library name (e.g libusb-1.0.so.0),
which is sufficient for ctypes to load the library, but not enough for
pyinstaller to find the file and pack it. This commit solves this problem
by using _resolveCtypesImports to figure out the full path to the file
pyusb loads.

Fixes #2633