Hooks: Guard pyusb-based discovery with an environment variable.
Add a check for an environment variable named
PYINSTALLER_USB_HOOK_SKIP_PYUSB_DISCOVERY. If the environment
variable is set and non-empty, the USB hook will skip pyusb-based
backend discovery and prefer the filesystem library search method
instead.

On some Linux machines (at least some Ubuntu 14.04 deployments),
unloading pyusb causes a segfault inside of libusb_exit() and crashes
the entire Python process. When this happens inside of the USB hook,
it crashes the PyInstaller hook, which crashes the entire PyInstaller
process.

There is no portable way to detect and gracefully recover from a
native segfault crash in Python. This approach is not ideal, but at
least provides a way for users to avoid crashes through
configuration.