building: util: Disallow empty DEST when adding binary or data files.
Instead, if an empty string is passed, PyInstaller will now terminate with a
useful message.

The destination-path `--add-data` and `--add-binary` had an undocumented
"feature": The destination-path (second element) could have been be an empty
string, in which case the target directory's name is the source directory's
basename.

The same was true for the `Analysis()` parameters `data` and `binaries`, but
here this was documented in the code - while not in the `Analysis()` API
documentation.

Rational

* The behavior is not documented in the manual and is not documented well in
  the code.

* The behavior is not obvious.

* This only works if the source is a directory or the source-glob only
  contains directories. As soon as the source is a file (of a glob resulting
  in a file), the empty string is not handled in the code and files will end
  up somewhere.

* The same can be achieved easily by just adding the directory basename as
  destination.

* This is a source of errors, see e.g. gh-3062.

Closes gh-3066.