Archive/crypt: Fix incompatibility with `pycryptodome`.
The `pycryptodome` library is a replacement for the (apparently
abandoned) `pycrypto` library, which is a bit more picky concerning
the `key` type. It isn't accepting `str` anymore and requires a byte
array to be passed to the C function. This is fixed by simply calling
encode on the string. This still works with the old `pycrypto`
library.