Update on hook-PyQt5.QtQml
Hooks:

Update on the hook-PyQt5.QtQml that ensure that:

  - Resources added in the application which do not use an rcc are found and added to the datas global variable.

  - The main Qml file is found and added to the bundled appliction.

  - Qml files imported within the qml source code are found and added.

  - Entire folders imported within the qml source is imported.

  - Hidden imports within the qml source code are added as well.

Mere inclusion of every qml file in the path/to/PyQt5/Qt/qml folder is costly.

Which means that the bundles app couldn't possible go below 15MB, even if you created a mere Hello world app with Qml.

So what has been done here is the search and mapping of the exact import called for in the application.

I am in no way promising a complete work, but the days are longer for me. And this is a complete work for those using Qt Quick2 and 15 other imports more. With the only two exceptions being the importation of the entire path/to/PyQt5/Qt/qml/Qt folder and a misnamed or so to speak Qt3D import statements. The later do not work at all.