compat: python library lookup on RHSCL python 3.6
The python 3.6 implementation on a RHEL 6 & 7 is packaged through
SCL (Sofware Collections).

It uses a non default implementation scheme where libraries are
installed in a chroot like environment (/opt/rh/rh-python36).

Content of /opt/rh/rh-python36/root/lib64
`
lrwxrwxrwx  libpython3.6m.so -> libpython3.6m.so.rh-python36-1.0
-rwxr-xr-x  libpython3.6m.so.rh-python36-1.0
-rwxr-xr-x  libpython3.so.rh-python36
`
This patch extend the current python library matching scheme on
unix with libpython%d.%dm.so to match libpython3.6m.so.
The symlink is followed and the libpython3.6m.so.rh-python36-1.0
library is embeded in the archive as expected.