depend: include anaconda path to libpython on linux
Anaconda is widely used. It is the recommeded way to install python and its modules.
PyInstaller works well with it in windows but did not consider searching for the ```libpython``` in the anaconda distrubution on linux which is most probably the case.
So here is a commit that solves the problem.