depend: improve Windows pylib detection
Allow pylib detection when VERSION.dll is a dependency of python.exe
rather than pythonXY.dll

If libpython is not detected uses the dependency checker to look for a
dll matching the pattern pythonXY.dll using the same method as
automated detection.

Checks after automated detection and only on windows systems.

Fixes #3942