Hooks: Add hook for ITK.
The Insight Segmentation and Registration Toolkit (ITK) is is an open-source,
cross-platform system that provides developers with an extensive suite of
software tools for image analysis [1]. ITK is written in C++ and wrapped in
Python. Its packages are available on PyPI for Linux, MacOS and Windows [2].
ITK's functionalities are completed with remote modules that are also
wrapped in Python and available on PyPI.

A sample of the ITK remote modules includes itk-vtkglue [3],
itk-isotropicwavelets [4], and itk-ultrasound [5].

Note that ITK's hook will only work if ITK is installed with pip. It will not
work when using ITK directly from its build tree.

[1] https://itk.org
[2] https://pypi.org/project/itk
[3] https://pypi.org/project/itk-vtkglue
[4] https://pypi.org/project/itk-isotropicwavelets
[5] https://pypi.org/project/itk-ultrasound